refactor: change class name
- 클래스명 변경
