+++ node-connection-backend/src/main/java/node/connection/hyperledger/indy/IndyUtils.java
@@ -1,9 +1,8 @@
-package node.connection.hyperledger;
+package node.connection.hyperledger.indy;
 
 import lombok.extern.slf4j.Slf4j;
 import node.connection._core.exception.ExceptionStatus;
 import node.connection._core.exception.server.ServerException;
-import node.connection.wallet.WalletConfigFactory;
 import org.hyperledger.indy.sdk.IndyException;
 import org.hyperledger.indy.sdk.pool.Pool;
 import org.hyperledger.indy.sdk.wallet.Wallet;
@@ -25,15 +24,20 @@ public class IndyUtils {
         this.walletConfigFactory = walletConfigFactory;
     }
 
-    public Pool initializeIndyConnectionPool() throws IndyException, ExecutionException, InterruptedException {
-        Pool.setProtocolVersion(2);
+    public Pool initializeIndyConnectionPool() {
         Pool pool = null;
 
         try {
+            Pool.setProtocolVersion(2);
             pool = Pool.openPoolLedger(POOL_NAME, "{}").get();
         } catch (Exception exception) {
-            Pool.createPoolLedgerConfig(POOL_NAME, POOL_CONFIG);
-            pool = Pool.openPoolLedger(POOL_NAME, "{}").get();
+            try {
+                log.info("Hyperledger Indy pool create using " + POOL_CONFIG);
+                Pool.createPoolLedgerConfig(POOL_NAME, POOL_CONFIG);
+                pool = Pool.openPoolLedger(POOL_NAME, "{}").get();
+            } catch (IndyException | ExecutionException | InterruptedException e) {
+                throw new ServerException(ExceptionStatus.INDY_INITIALIZATION_ERROR);
+            }
         }
 
         log.info("Hyperledger Indy Pool created and opened successfully.");
